Optimize parsing a number

+++ src/main/java/szewek/flux/gui/SignalControllerScreen.java
@@ -127,7 +127,7 @@ public class SignalControllerScreen extends ContainerScreen<SignalControllerCont
 			return false;
 		}
 		try {
-			short st = Short.parseShort(txt);
+			int st = Integer.parseInt(txt);
 			if (st < 0 || st > 255) {
 				return false;
 			}
